% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dong_cell_anno.R
\name{dong_cell_anno}
\alias{dong_cell_anno}
\title{Title}
\usage{
dong_cell_anno(
  sce,
  sce_ref = NULL,
  name = "Model1",
  merge_seurat_cluster = T,
  name_merge_seurat = "scpred_seurat",
  mini_cluster = 50,
  threshold = 0.65,
  source = "win",
  path_ref = NULL
)
}
\arguments{
\item{path_ref}{}
}
\description{
Title
}
