% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dong_dimplot.R
\name{dong_dimplot}
\alias{dong_dimplot}
\title{Modified_Dimplot}
\usage{
dong_dimplot(
  sce,
  reduction = "umap",
  groups = "orig.ident",
  split.by = NULL,
  label = T,
  label.size = 5,
  pt.size = 0.5,
  cols = "normal",
  seed = 123,
  palette = 1,
  show_col = FALSE,
  width = 9.2,
  height = 8.2,
  w_index = 7,
  w_add = 2,
  max_category = 18,
  show_plot = T,
  path = NULL,
  index = 1,
  theme = "classic",
  legend.position = "right",
  legend.direction = "horizontal",
  legend.size = 0.25,
  plot_title_size = 2,
  axis_title_size = 2,
  axis_text_size = 10,
  axis_angle = 0,
  hjust = 0.5,
  save_plot = TRUE,
  fig.type = "pdf"
)
}
\arguments{
\item{sce}{seurat object}

\item{reduction}{default is umap, other options: tsne, pca}

\item{groups}{colors of group}

\item{split.by}{variables used to split the figures}

\item{label}{true or false}

\item{label.size}{size of label}

\item{pt.size}{size of point}

\item{cols}{users can define the cols manualy}

\item{seed}{default is 54321}

\item{show_col}{default is FALSE}

\item{width}{width of figure}

\item{height}{height of figure}

\item{fig.type}{}
}
\description{
Modified_Dimplot
}
