% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dong_find_markers.R
\name{dong_find_markers}
\alias{dong_find_markers}
\title{dong_find_markers}
\usage{
dong_find_markers(
  sce,
  assay = NULL,
  slot = "scale.data",
  group = NULL,
  verbose = FALSE,
  fig.type = "pdf",
  pt.size = 1,
  cols = "normal",
  seed = 123,
  show_col = F,
  palette = 1,
  show_genes = 10,
  show_genes_pheatmap = 5,
  logfc.threshold = 0.15,
  test.use = "wilcox",
  only.pos = TRUE,
  feature_type = "gene",
  enrich_cutoff_logfc = 0.2,
  enrich_cutoff_padj = 0.05,
  hwidth = 10,
  hheight = NULL,
  show_features = 8,
  show_plot = T,
  path = NULL,
  character_limit = 50,
  recluster = FALSE,
  assay_for_recluster,
  dims_for_recluster = 20,
  resolution_for_recluster = 0.2
)
}
\arguments{
\item{palette}{}
}
\description{
dong_find_markers
}
\examples{
violin-plot-and-feature-plot-of-each-clusters-----------------------------
}
