% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dong_heatmap.R
\name{dong_heatmap}
\alias{dong_heatmap}
\title{Title}
\usage{
dong_heatmap(
  sce,
  group,
  feas = NULL,
  assay = NULL,
  slot = "scale.data",
  cols = "normal",
  palette = 1,
  palette.heatmap = 1,
  path = NULL,
  prefix = 0,
  show_plot = TRUE,
  show_col = FALSE,
  fig.type = "pdf"
)
}
\arguments{
\item{sce}{}

\item{group}{}

\item{feas}{}

\item{assay}{}

\item{slot}{}

\item{cols}{}

\item{palette}{}

\item{palette.heatmap}{}

\item{path}{}

\item{prefix}{}

\item{show_plot}{}

\item{show_col}{}

\item{fig.type}{}
}
\description{
Title
}
