% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doublet_detect.R
\name{doublet_detect}
\alias{doublet_detect}
\title{Integration of  doublet detection}
\usage{
doublet_detect(
  sce = NULL,
  already_normalized = FALSE,
  eset = NULL,
  file_type = "10X",
  data_path = NULL,
  project = NULL,
  index = 1,
  propotion = 0.025,
  check_data = TRUE,
  filter_data = TRUE,
  minFeature = 2000,
  minCount = 1000,
  percent.mt = 20,
  method = "doubletfinder",
  cores = 1,
  save_path = NULL
)
}
\arguments{
\item{sce}{Seurat object, default is NULL}

\item{already_normalized}{if TRUE, input data has been normalized}

\item{eset}{Raw data or normalized data}

\item{file_type}{If file is from 10X, data_path must be provide}

\item{data_path}{Absolute path of sparse data matrices provided by 10X genomics}

\item{project}{Project name for the Seurat object, such as TNBC. If save_path is NULL, the project name will be part of folder name of save_path. default is NULL}

\item{index}{Prefix of the file name for saving}

\item{propotion}{Doublet formation rate,  default is 0.025}

\item{check_data}{If TRUE, quality control will be done}

\item{filter_data}{If TRUE, cells will be filtered by aforementioned criterion}

\item{minFeature}{Minimum features of cell}

\item{minCount}{Minimum counts of cell}

\item{percent.mt}{Maximum of percentage of mitochondrial gens}

\item{method}{Methods for doublet detection, choose from `doubletfinder` and `scds`}

\item{cores}{Number of cores to use for parallelization, default is 1}

\item{save_path}{Path of the output saving directory}
}
\value{
Seurat object with updated metadata including pANN and doublet classifications when `method= "doubletfinder"`;
Seurat object with updated metadata including doublet scores by `cxds`, `bcds` and `cxds_bcds_hybrid`, and doublet classifications when `method= "scds"`
}
\description{
Here, we provide two doublet-detection methods:`doubletFinder{DoubletFinder}` and `scds`,
since the DoubletFinder method has the best detection accuracy, and the cxds method has the highest computational efficiency[1]
You can use `help(doubletFinder)` and `help(scds)` for more details.
}
\examples{
data("pbmc_small")
doublet_detect( sce=pbmc_small)
}
