% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_cluster_degs.R
\name{enrich_cluster_degs}
\alias{enrich_cluster_degs}
\title{enrich_cluster_degs}
\usage{
enrich_cluster_degs(
  sce.markers,
  cutoff_foldchange = 0.5,
  cutoff_p_adj = 0.01,
  methods = c("enrichGO", "enrichKEGG"),
  cluster = "cluster",
  gene = "gene",
  p.adj = "p_val_adj",
  logfc = "avg_log2FC",
  showCategory = 12,
  path = NULL,
  cols = c("#b3eebe", "#46bac2", "#371ea3"),
  index = NULL,
  width = 9.36,
  height = 14.3
)
}
\arguments{
\item{sce.markers}{result of find_markers}

\item{cutoff_foldchange}{default is 0.5}

\item{cutoff_p_adj}{"enrichGO", "enrichKEGG"}

\item{showCategory}{number of terms to show}

\item{index}{default is NULL}
}
\description{
enrich_cluster_degs
}
