% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_sc_data.R
\name{extract_sc_data}
\alias{extract_sc_data}
\title{Title}
\usage{
extract_sc_data(
  sce,
  vars = NULL,
  assay,
  slot = "scale.data",
  combine_meta_data = TRUE
)
}
\arguments{
\item{combine_meta_data}{}
}
\description{
Title
}
