% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_subcluster_gene.R
\name{find_subcluster_gene}
\alias{find_subcluster_gene}
\title{Title}
\usage{
find_subcluster_gene(
  sce,
  assay = "integrated",
  col_celltype = "celltypes",
  col_sub_celltype = "scpred_seurat",
  cols = "normal",
  seed = 123,
  palette = 1,
  show_col = F,
  genes_for_deg = NULL,
  min_cell_count = 30,
  no_limitation_celltypes = "Epithelial cells",
  index = 1,
  path = "4-Diff-degs-of-subclusters",
  show_feas_pheatmap = 8,
  remove_other_celltypes = TRUE,
  find_cluster_sig = FALSE
)
}
\arguments{
\item{sce}{}

\item{assay}{}

\item{col_celltype}{}

\item{col_sub_celltype}{}

\item{cols}{}

\item{seed}{}

\item{palette}{}

\item{show_col}{}

\item{genes_for_deg}{}

\item{min_cell_count}{}

\item{no_limitation_celltypes}{}

\item{index}{}

\item{path}{}

\item{show_feas_pheatmap}{}

\item{remove_other_celltypes}{}

\item{find_cluster_sig}{}
}
\description{
Title
}
\examples{
如果目标变量没有亚组直接跳过此次循环====================
}
