% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_subclusters.R
\name{find_subclusters}
\alias{find_subclusters}
\title{Sub-cluster determination}
\usage{
find_subclusters(
  sce,
  celltype,
  combined_subclusters = "celltype_subcluster",
  graph.name = NULL,
  resolution = 0.3,
  delete_new_variable = T,
  filtering_mini_cell = TRUE,
  min_cell_count = 50,
  ignore_cell_prefix = "null"
)
}
\arguments{
\item{sce}{Seurat object}

\item{celltype}{Name of the metadata column to group cells by}

\item{combined_subclusters}{New name of a metadata column, dedault is "celltype_subcluster"}

\item{graph.name}{Name of graph to use for the clustering algorithm}

\item{resolution}{Value of the resolution parameter, use a value above (below) 1.0 if you want to obtain a larger (smaller) number of communities}

\item{delete_new_variable}{Dedault is TRUE}

\item{filtering_mini_cell}{Whether to filter out the clusters with few cells, dedault is TRUE}

\item{min_cell_count}{Minimum number of cell counts of each cluster, dedault is 50}

\item{ignore_cell_prefix}{A character vector corresponding to cell types with no cell count limitation}
}
\value{
Suerat object with updated metadata containing subgroups of cells
}
\description{
Identify sub-clusters for each cluster of cells
}
\examples{
## Not RUN
data("pbmc_small")
sce<-find_subclusters(sce=pbmc_small, celltype= "RNA_snn_res.1")
## Not RUN
Batch search for subgroups of cells of a group of cell =================================
Integrating the above cell subgroups information into one column======combined_column==================================
Removal of some discrete cells
}
