% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scsig_cell_anno.R
\name{format_sig_from_df}
\alias{format_sig_from_df}
\title{Title}
\usage{
format_sig_from_df(
  deg,
  cluster = "cluster",
  gene = "gene",
  avg_log2FC = "avg_log2FC",
  n = 100
)
}
\arguments{
\item{avg_log2FC}{}
}
\description{
Title
}
