% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sctype_score.R
\name{gene_sets_prepare}
\alias{gene_sets_prepare}
\title{gene set prepare
Written by Aleksandr Ianevski <aleksandr.ianevski@helsinki.fi>, June 2021
Modified by IOBR organization <iobr2019@163.com>, Oct 2022}
\usage{
gene_sets_prepare(
  gs_data = NULL,
  path_to_db_file = NULL,
  cell_type = NULL,
  tissue_type = NULL,
  cell_subset = NULL,
  study = NULL
)
}
\arguments{
\item{gs_data}{user can provide gene set data manually}

\item{path_to_db_file}{- DB file with cell types}

\item{cell_type}{default is null}

\item{tissue_type}{default is null}

\item{cell_subset}{default is null}

\item{study}{}
}
\description{
gene set prepare
Written by Aleksandr Ianevski <aleksandr.ianevski@helsinki.fi>, June 2021
Modified by IOBR organization <iobr2019@163.com>, Oct 2022
}
