% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cols.R
\name{get_cols}
\alias{get_cols}
\title{Title}
\usage{
get_cols(cols = "normal", palette = 1, show_col = T, seed = 123)
}
\arguments{
\item{cols}{}

\item{palette}{}

\item{show_col}{}

\item{seed}{}
}
\description{
Title
}
