% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sctype_score.R
\name{get_gene_sets}
\alias{get_gene_sets}
\title{Get gene sets from DB, or modified gene sets data}
\usage{
get_gene_sets(
  gs = NULL,
  path_to_db_file = NULL,
  cell_type = NULL,
  tissue_type = NULL,
  cell_subset = NULL
)
}
\arguments{
\item{gs}{user can provide gene set file manually}

\item{path_to_db_file}{- DB file with cell types}

\item{cell_type}{default is null}

\item{tissue_type}{default is null}

\item{cell_subset}{default is null}
}
\description{
Get gene sets from DB, or modified gene sets data
}
