% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycopykat.R
\name{mycopykat}
\alias{mycopykat}
\title{copycat_main_function}
\usage{
mycopykat(
  rawmat,
  id.type = "S",
  cell.line = "no",
  ngene.chr = 5,
  LOW.DR = 0.05,
  UP.DR = 0.1,
  win.size = 25,
  norm.cell.names = "",
  KS.cut = 0.1,
  sam.name = "E",
  distance = "euclidean",
  n.cores = 1,
  save_path = NULL
)
}
\arguments{
\item{rawmat}{raw data matrix; genes in rows; cell names in columns.}

\item{id.type}{gene id type: Symbol or Ensemble.}

\item{cell.line}{if the data are from pure cell line,put "yes"; if cellline data are a mixture of tumor and normal cells, still put "no".}

\item{ngene.chr}{minimal number of genes per chromosome for cell filtering.}

\item{LOW.DR}{minimal population fractions of genes for smoothing.}

\item{UP.DR}{minimal population fractions of genes for segmentation.}

\item{win.size}{minimal window sizes for segmentation.}

\item{norm.cell.names}{a vector of normal cell names.}

\item{KS.cut}{segmentation parameters, input 0 to 1; larger looser criteria.}

\item{sam.name}{sample name.}

\item{distance}{distance methods include euclidean, and correlation coverted distance include pearson and spearman.}

\item{n.cores}{number of cores for parallel computing.}

\item{save_path}{absolute path to save result}
}
\value{
aneuploid/diploid prediction results; 2.CNA results in 220KB windows; 3.heatmap; 4.hclustering object.
}
\description{
copycat_main_function
}
