% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_strata_cells.R
\name{random_strata_cells}
\alias{random_strata_cells}
\title{randomly strata cells}
\usage{
random_strata_cells(
  input,
  group,
  propotion = 0.1,
  minum_count_include = 300,
  minum_count = 200,
  max_count = 1000,
  sub_cluster = NULL,
  cell_type = NULL
)
}
\arguments{
\item{input}{cell annotation with cell id}

\item{group}{annotation of cell type}

\item{propotion}{range between 0 to 1}

\item{minum_count}{minimum count of each cell type}

\item{sub_cluster}{default is NULL, if this parameter was set}

\item{cell_type}{default is NULL, The default identity of Seurat object will be used}
}
\description{
randomly strata cells
}
