% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_fea_plot.R
\name{sc_fea_plot}
\alias{sc_fea_plot}
\title{single cell features plots}
\usage{
sc_fea_plot(
  sce,
  group = NULL,
  sub_group = NULL,
  target = NULL,
  remove_other_celltypes = FALSE,
  min_cell_count = 10,
  show_label = TRUE,
  split_by = NULL,
  assay = NULL,
  slot = "scale.data",
  variables = NULL,
  show_variables = 10,
  dims = c("umap", "tsne"),
  pt.size = 1,
  cols = "normal",
  palette = 1,
  show_cols = FALSE,
  palette.heatmap = 1,
  path = NULL,
  index = NULL,
  show_box_pvalue = T,
  fig.type = "pdf",
  show_plot = FALSE,
  dims_for_recluster = 30
)
}
\arguments{
\item{sce}{Seurat object after umap and tsne dimensionality reduction}

\item{group}{Name of a metadata column to group cells by}

\item{sub_group}{Name of a metadata column to re-group a subset of cells by}

\item{target}{Vector of specific cell identity classes in `group`}

\item{remove_other_celltypes}{Default is False}

\item{min_cell_count}{Minimal cell counts in each cluster, default is 10}

\item{show_label}{If TRUE, labels of cell clusters will be shown in featurePlot}

\item{split_by}{Name of a metadata column to split plot by}

\item{assay}{Assay to pull data from}

\item{slot}{Data slot to use, options: scale.data, data, counts}

\item{variables}{Vector of variables to show in plot}

\item{show_variables}{Number of variables to show in plot, default is 10}

\item{dims}{Vector of dimensionality reduction methods, such as 'umap' and 'tsne'}

\item{pt.size}{Point size of 'FeaturePlot', default is 1}

\item{cols}{Vector of colors, users can define the cols manually.  This may also be a single character, such as normal and random,  to a palette as specified by `palettes(){IOBR}`
See [palettes](http://127.0.0.1:60491/help/library/IOBR/html/palettes.html) for details}

\item{palette}{Numeric value corresponding with color palette. Default is 1, other options: 2, 3, 4}

\item{palette.heatmap}{Numeric value corresponding with colors to use for n-colour gradient in heatmap, default is 1, other options: 2, 3, 4, 5, 6}

\item{path}{Path of the output saving directory}

\item{index}{Prefix of file name for saving}

\item{show_box_pvalue}{If TRUE, paired wised statistical p-value will be shown in boxplot}

\item{fig.type}{Format of plot saving, such as pdf and png}

\item{show_plot}{If TRUE, pheatmap of all selected variables will be shown in the Rstudio}

\item{dims_for_recluster}{Number of PCs for re-clustering, default is 30}
}
\value{
Data frame with cells as rows and features as columns
}
\description{
Visualize features for specific cell identity classes. First select features with significant differences from the input features.
Then use a variety of plot types to visualize above features for specific cell identity classes, including box plots, umap plots, tsne plots,
violin plots, and heatmap.
}
\examples{
data("pbmc_small")
pbmc_small
# Run UMAP map on first 20 PCs
pbmc_small <- RunUMAP(object = pbmc_small, dims = 1:10)
#Choose features
variables<-c("PPBP", "IGLL5", "VDAC3", "CD1C", "AKR1C3", "PF4", "MYL9", "GNLY", "TREML1", "CA2")
res<-sc_fea_plot(sce= pbmc_small, variables= variables, group= "RNA_snn_res.1", assay="RNA")
choose cell identity classes-------------------------
}
\author{
Dongqiang Zeng
}
