% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_sig_plot.R
\name{sc_sig_plot}
\alias{sc_sig_plot}
\title{single cell signature score plots}
\usage{
sc_sig_plot(
  sce,
  group = NULL,
  show_label = TRUE,
  split_by = NULL,
  assay = NULL,
  slot = "scale.data",
  variables,
  show_variables = 10,
  dims = c("umap", "tsne"),
  pt.size = 1,
  cols = NULL,
  palette = NULL,
  path = NULL,
  index = NULL,
  show_box_pvalue = T,
  fig.type = "pdf",
  show_plot = FALSE
)
}
\arguments{
\item{sce}{seurat object}

\item{group}{cell identity}

\item{show_label}{default is true: featurePlot will show labels of cell clsuters}

\item{assay}{assay that stored variables}

\item{slot}{data type used to estimation, options: scale.data, data, counts}

\item{variables}{targets that users want to draw plots}

\item{show_variables}{Maximum number of variables displayed}

\item{dims}{default is 'umap' and 'tsne'}

\item{pt.size}{defalut is 1, point size of FeaturePlot}

\item{cols}{specifying colors of cell clusters}

\item{palette}{default is NULL, options: jama, jco, aaas, set2, and so on, functions derived from IOBR::palettes()}

\item{path}{default is "0-",group,"-sig-plot", users can specifying the path}

\item{index}{default is null}

\item{show_box_pvalue}{default is true, if true, boxplot will show the paired wised statistical p-value}

\item{fig.type}{default is pdf, other options: png}

\item{show_plot}{if true, pheatmap of all selected variables will be shown in the R studio}
}
\description{
single cell signature score plots
}
\examples{

选择需要分析的变量------labels不多的-------------------------
}
