% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_type_anno.R
\name{sc_type_anno}
\alias{sc_type_anno}
\title{single cell annotation by scType}
\usage{
sc_type_anno(
  sce,
  gs = NULL,
  tissue_type = NULL,
  cell_type = "base",
  cell_subset = NULL,
  study = NULL,
  assay = NULL,
  slot = "scale.data",
  scale = NULL,
  cluster = "seurat_clusters",
  cols = "normal",
  palette = 1,
  show_col = T,
  seed = 123,
  point.size = 1.5,
  reduction = "umap",
  db_ = "ScTyperDB-merged.xlsx",
  gene_names_to_uppercase = TRUE
)
}
\arguments{
\item{sce}{seurat object}

\item{gs}{default is null, user can provide gene set file manually}

\item{tissue_type}{default is null = base}

\item{cell_type}{default is null,}

\item{cell_subset}{default is null}

\item{assay}{default is RNA}

\item{slot}{default is scaled data}

\item{scale}{default is NULL}

\item{cluster}{cluster used to annotation}

\item{cols}{'random', 'normal', or colors}

\item{palette}{options: 1, 2}

\item{show_col}{default is TRUE}

\item{seed}{default is 123 if cols = 'random'}

\item{point.size}{default is 1.5}

\item{reduction}{default is umap}

\item{db_}{default is ScTypeDB_full.xlsx, deposited in data}
}
\description{
single cell annotation by scType
}
\examples{
tnbc<-sc_type_anno(sce = sce, tissue = "Immune system")
}
