% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scpred_cell_anno.R
\name{scpred_cell_anno}
\alias{scpred_cell_anno}
\title{single cell annotation by scPred}
\usage{
scpred_cell_anno(
  sce,
  sce_ref = NULL,
  name = "Model1",
  merge_seurat_cluster = T,
  name_merge_seurat = "scpred_seurat",
  mini_cluster = 50,
  threshold = 0.65,
  source = "win",
  path_ref = NULL
)
}
\arguments{
\item{sce}{A seurat object containing cells to be classified}

\item{sce_ref}{A Seurat object with trained model(s) using scPred or a scPred object}

\item{name}{A character used to name a new column of metadata containing cell type annotation by `scPred`}

\item{merge_seurat_cluster}{Whether to merge prediction results by`scPred` with clustering results}

\item{name_merge_seurat}{A character used to name a new column of metadata containing the results of `merge_seurat_cluster`}

\item{mini_cluster}{Minimal cell counts of the cluster}

\item{threshold}{Threshold used for probabilities to classify cells into classes. All cells below this threshold value will be labels as "unassigned". In the case of binary classification (two cell tyoes), a threshold of 0.5 will force all cells to be classified to any of the two cell types. For multi-class classification, if there's no probability higher than the threshold associated to a cell type, this will be labelled as "unassigned"}

\item{source}{Character string related to computer system name choose from win and linux}

\item{path_ref}{The path or the name of the file where the `sce_ref` is read from}
}
\value{
Seurat object with updated metadata containing cell type annotation by `scPred`
}
\description{
Cell type annotation using a combination of unbiased feature selection from a reduced-dimension space, and machine-learning probability-based prediction method.
 Refer to [scPred](https://genomebiology.biomedcentral.com/articles/10.1186/s13059-019-1862-5)
}
