% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scsig_cell_anno.R
\name{scsig_cell_anno}
\alias{scsig_cell_anno}
\title{Title}
\usage{
scsig_cell_anno(
  sce,
  gs = NULL,
  gs_data = NULL,
  deg = NULL,
  top_n_deg = 100,
  method = c("pca", "sctyper", "model", "censu"),
  assay = NULL,
  slot = "scale.data",
  min.feature = 10,
  scale = NULL,
  cluster = "seurat_clusters",
  model_scpred = NULL,
  merge_seurat_cluster = FALSE,
  mini_cluster = 50,
  threshold = 0.65,
  source = "win",
  path_model_scpred = NULL,
  tissue_type = NULL,
  gs_sctyper = NULL,
  db_ = "ScTyperDB-merged.xlsx",
  db_path = NULL,
  cell_type = "base",
  cell_subset = NULL,
  study = NULL,
  gene_names_to_uppercase = TRUE,
  cols = "normal",
  palette = 1,
  show_col = FALSE,
  seed = 123,
  point.size = 1.5,
  reduction = "umap",
  show_plot = TRUE,
  save_plot = TRUE,
  width = 12,
  height = 5,
  path = NULL,
  fig.type = "pdf"
)
}
\arguments{
\item{cols}{Vector of colors, users can define the cols manually.  This may also be a single character, such as normal and random,  to a palette as specified by `palettes(){IOBR}`
See [palettes](http://127.0.0.1:60491/help/library/IOBR/html/palettes.html) for details}

\item{palette}{Numeric value corresponding with color palette. Default is 1, other options: 2, 3, 4}

\item{show_col}{Whether to show color palettes}

\item{seed}{Seed of the random number generator, default is 123. The parameter works when cols ="random"}

\item{point.size}{Size of point, default is 1.5}

\item{reduction}{Which dimensionality reduction to use. If not specified, first searches for umap, then tsne, then pca}

\item{fig.type}{}
}
\description{
Title
}
