% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scsig_dimplot.R
\name{scsig_dimplot}
\alias{scsig_dimplot}
\title{Modified dimensional reduction plot}
\usage{
scsig_dimplot(
  sce,
  reduction = "umap",
  groups = "orig.ident",
  split.by = NULL,
  label = T,
  label.size = 5,
  pt.size = 0.5,
  cols = "normal",
  seed = 123,
  palette = 1,
  show_col = FALSE,
  width = 9.2,
  height = 8.2,
  w_index = 7,
  w_add = 2,
  max_category = 18,
  show_plot = T,
  path = NULL,
  index = 1,
  theme = "classic",
  legend.position = "right",
  legend.direction = "horizontal",
  legend.size = 0.25,
  plot_title_size = 2,
  axis_title_size = 2,
  axis_text_size = 10,
  axis_angle = 0,
  hjust = 0.5,
  save_plot = TRUE,
  fig.type = "pdf"
)
}
\arguments{
\item{sce}{Seurat object}

\item{reduction}{Which dimensionality reduction to use. If NULL, first searches for umap, then tsne, then pca.}

\item{groups}{Name of one or more metadata columns to group (color) cells by (for example, orig.ident); pass 'ident' to group by identity class}

\item{split.by}{Name of a metadata column to split plot by}

\item{label}{Whether to label the clusters}

\item{label.size}{Size of label}

\item{pt.size}{Size of point}

\item{cols}{Vector of colors, each color corresponds to an identity class. This may also be a single character, such as normal and random, to a palette as specified by `IOBR::palettes`. See `palettes` for details}

\item{seed}{Seed of the random number generator, default is 123. The parameter works when cols ="random"}

\item{palette}{Numeric value corresponding with color palette. Default is 1, other options: 2, 3, 4}

\item{show_col}{Whether to display the palettes}

\item{width}{Width of plot when saving}

\item{height}{Height of plot when saving}

\item{w_index}{Numeric value corresponding to `width`. The parameter works when `split.by` not NULL}

\item{w_add}{Numeric value corresponding to `width`. The parameter can be used to increase `width` when necessary}

\item{max_category}{Maximal number of groups}

\item{show_plot}{Whether to display the plot}

\item{path}{Path of the output saving directory}

\item{index}{Index number of folder name}

\item{theme}{Theme of plot.}

\item{legend.position}{Position of legend}

\item{legend.direction}{Direction of legend}

\item{legend.size}{Font size of legend}

\item{plot_title_size}{Title size}

\item{axis_title_size}{Axis title size}

\item{axis_text_size}{Axis text size}

\item{axis_angle}{Axis angle}

\item{hjust}{Horizontal justification}

\item{save_plot}{Whether to save the plot}

\item{fig.type}{Format of plot saving, such as pdf and png}
}
\value{
A list of ggplot objects
}
\description{
Modified dimensional reduction plot
}
\examples{
data("pbmc_small")
scsig_dimplot(object = pbmc_small)
}
\author{
Dongqiang Zeng
}
