% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scsig_find_markers.R
\name{scsig_find_markers}
\alias{scsig_find_markers}
\title{Find and show gene expression markers for all identity classes}
\usage{
scsig_find_markers(
  sce,
  assay = NULL,
  adjust_assay = FALSE,
  slot = "scale.data",
  feas = NULL,
  group = NULL,
  verbose = FALSE,
  fig.type = "pdf",
  pt.size = 1,
  cols = "normal",
  seed = 123,
  show_col = F,
  palette = 1,
  show_genes = 10,
  show_genes_pheatmap = 5,
  logfc.threshold = 0.15,
  test.use = "wilcox",
  only.pos = TRUE,
  feature_type = "gene",
  enrich_cutoff_logfc = 0.2,
  enrich_cutoff_padj = 0.05,
  hwidth = 10,
  hheight = NULL,
  show_features = 8,
  show_plot = T,
  path = NULL,
  character_limit = 50,
  re_scale_tsne_umap = FALSE,
  recluster = FALSE,
  assay_for_recluster = NULL,
  dims_for_recluster = 20,
  group_after_recluster = "default",
  resolution_for_recluster = 0.2
)
}
\arguments{
\item{sce}{Seurat object}

\item{assay}{Assay to use in differential expression testing}

\item{adjust_assay}{default is FALSE}

\item{slot}{Slot to pull data from; note that if test.use is "negbinom", "poisson", or "DESeq2", slot will be set to "counts"}

\item{feas}{Genes to test. Default is to use all genes}

\item{group}{A vector of variables to group cells}

\item{verbose}{Whether to print a progress bar once expression testing begins}

\item{fig.type}{Format of plot saving, such as pdf and png}

\item{pt.size}{Size of point}

\item{cols}{Vector of colors, each color corresponds to an identity class. This may also be a single character, such as normal and random,  to a palette as specified by `IOBR::palettes`. See `palettes` for details}

\item{seed}{Seed of the random number generator, default is 123. The parameter works when cols ="random"}

\item{show_col}{Whether to display the palettes}

\item{palette}{Numeric value corresponding with color palette. Default is 1, other options: 2, 3, 4}

\item{show_genes}{Number of genes displayed in the heatmap drawn by `DoHeatmap()`, default is 10}

\item{show_genes_pheatmap}{Number of genes displayed in the heatmap drawn by `pheatmap_average()`, default is 5}

\item{logfc.threshold}{Limit testing to genes which show, on average, at least X-fold difference (log-scale) between the two groups of cells. Default is 0.15 Increasing logfc.threshold speeds up the function, but can miss weaker signals}

\item{test.use}{Denotes which test to use. Available options are: "wilcox"(default), "bimod", "roc", "t", "nr", "negbinom", "poisson", "LR", "MAST", "DESeq2".}

\item{only.pos}{Only return positive markers, default is T}

\item{feature_type}{Feature type, default is gene, otherwise the enrichment analysis will not be done.}

\item{enrich_cutoff_logfc}{Cutoff for filtering gene signatures which show on average, less than X-fold difference (log-scale) between the two groups of cells for enrichment analysis}

\item{enrich_cutoff_padj}{Cutoff for filtering genes which adjust P values are < n}

\item{hwidth}{width of plot when saving}

\item{hheight}{Height of plot when saving}

\item{show_features}{Number of genes displayed in the violin plot, default is 8}

\item{show_plot}{Whether to display the plot}

\item{path}{Path of the output saving directory}

\item{character_limit}{Limit length of character, default is 50}

\item{re_scale_tsne_umap}{if TRUE, dimension reduction will be proceeded in new data}

\item{recluster}{default is FALSE}

\item{assay_for_recluster}{Name of assay to use, such as RNA,SCT,integrated}

\item{dims_for_recluster}{Number of PCs to compute and store for re-clustering}

\item{group_after_recluster}{if TRUE, finding cluster  will be proceeded in new data}

\item{resolution_for_recluster}{Value of the resolution parameter for re-clustering}
}
\description{
This function can be used to find markers (differentially expressed genes) for each of the identity classes in a dataset
based on `FindAllMarkers(){Seurat}`. Then, use `DoHeatmap(){Seurat}` and `pheatmap_average()`to draw a heatmap
and `VlnPlot(){Seurat}` to draw a violin plot of single cell feature expression.
Besides, `enrich_cluster_degs()` is used to perform enrichment analysis.
}
\examples{
data("pbmc_small")
scsig_find_markers(sce = pbmc_small)
violin-plot-and-feature-plot-of-each-clusters-----------------------------
}
\author{
Dongqiang Zeng
}
