% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scsig_heatmap.R
\name{scsig_heatmap}
\alias{scsig_heatmap}
\title{Modified feature expression heatmap}
\usage{
scsig_heatmap(
  sce,
  group,
  feas = NULL,
  assay = NULL,
  slot = "scale.data",
  disp.min = -2.5,
  disp.max = 5,
  cols = "normal",
  palette = 1,
  palette.heatmap = 1,
  path = NULL,
  prefix = 0,
  show_plot = TRUE,
  show_col = FALSE,
  fig.type = "pdf"
)
}
\arguments{
\item{sce}{Seurat object}

\item{group}{A vector of variables to group cells by; pass 'ident' to group by cell identity classes}

\item{feas}{A vector of features to plot, defaults to `VariableFeatures(object = object)`}

\item{assay}{Assay to pull from}

\item{slot}{Data slot to use, choose from 'raw.data', 'data', or 'scale.data'}

\item{disp.min}{Minimum display value (all values below are clipped)}

\item{disp.max}{Maximum display value (all values above are clipped); defaults to 2.5 if slot is 'scale.data', 6 otherwise}

\item{cols}{Vector of colors, each color corresponds to an identity class. This may also be a single character, such as normal and random,  to a palette as specified by `IOBR::palettes`. See `palettes` for details}

\item{palette}{Numeric value corresponding with colors to use for the color bar. Default is 1, other options: 2, 3, 4}

\item{palette.heatmap}{Numeric value corresponding with colors  to use for n-colour gradient, default is 1, other options: 2, 3, 4, 5, 6}

\item{path}{Path of the output saving directory}

\item{prefix}{Prefix of the file name}

\item{show_plot}{Whether to show the plot}

\item{show_col}{Whether to show color palettes}

\item{fig.type}{Format of plot saving, such as pdf and png}
}
\value{
a list of ggplot objects
}
\description{
The function is a modified version of `DoHeatmap{Seurat}`.
}
\examples{
data("pbmc_small")
scsig_heatmap(sce = pbmc_small,group= "RNA_snn_res.1")
}
