% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sctype_score.R
\name{sctype_score}
\alias{sctype_score}
\title{Title}
\usage{
sctype_score(
  scRNAseqData,
  scaled = !0,
  gs_sctyper,
  gs_sctyper2 = NULL,
  gene_names_to_uppercase = !0,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
