% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sctyper_anno.R
\name{sctyper_anno}
\alias{sctyper_anno}
\title{Cell type annotation by scType}
\usage{
sctyper_anno(
  sce,
  gs_sctyper = NULL,
  tissue_type = NULL,
  cell_type = "base",
  cell_subset = NULL,
  study = NULL,
  assay = NULL,
  slot = "scale.data",
  scale = NULL,
  cluster = "seurat_clusters",
  db_ = "ScTyperDB-merged.xlsx",
  db_path = NULL,
  gene_names_to_uppercase = TRUE
)
}
\arguments{
\item{sce}{Seurat object}

\item{gs_sctyper}{Default is null, user can provide gene set file manually}

\item{tissue_type}{Default is NULL}

\item{cell_type}{Cell types choosed from "base", "epithelial", "myeloid", "tcell", "bcell", "fibroblast" and "endothelial", default is "base"}

\item{cell_subset}{Default is NULL}

\item{assay}{Assay to pull from, e.g. RNA, SCT, integrated}

\item{slot}{Data slot to use, choose from 'counts', 'data', or 'scale.data'}

\item{scale}{Whether the matrix is scaled，default is NULL}

\item{cluster}{A vector of variables to group cells by}

\item{db_}{Database of manually collected cell type annotation, default is "ScTypeDB_full.xlsx" deposited in data}

\item{point.size}{Size of point, default is 1.5}

\item{cols}{Vector of colors, users can define the cols manually.  This may also be a single character, such as normal and random,  to a palette as specified by `palettes(){IOBR}`
See [palettes](http://127.0.0.1:60491/help/library/IOBR/html/palettes.html) for details}

\item{palette}{Numeric value corresponding with color palette. Default is 1, other options: 2, 3, 4}

\item{show_col}{Whether to show color palettes}

\item{seed}{Seed of the random number generator, default is 123. The parameter works when cols ="random"}

\item{reduction}{Which dimensionality reduction to use. If not specified, first searches for umap, then tsne, then pca}
}
\value{
Seurat object with updated metadata containing cell type annotation
}
\description{
Cell type annotation using specific marker combinations from single-cell transcriptomic data
refer to [ScType](https://www.nature.com/articles/s41467-022-28803-w)
}
\examples{
tnbc<-sctyper_anno(sce = sce, tissue = "Immune system")
}
