% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unassign_cell.R
\name{unassign_cell}
\alias{unassign_cell}
\title{Title}
\usage{
unassign_cell(
  sce,
  cluster,
  ignore_cell_prefix = NULL,
  min_cell_count = 30,
  new_col = NULL,
  delete_unassigned = FALSE,
  return_meta_data = FALSE,
  alread_unassigned = "unassigned",
  minimal_cell = 10
)
}
\arguments{
\item{minimal_cell}{}
}
\description{
Title
}
