% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doublet_detect.R
\name{doublet_detect}
\alias{doublet_detect}
\title{Integration of  doublet detection}
\usage{
doublet_detect(
  sce = NULL,
  already_normalized = FALSE,
  eset = NULL,
  file_type = "10X",
  data_path = NULL,
  project = NULL,
  index = 1,
  propotion = 0.025,
  check_data = TRUE,
  filter_data = TRUE,
  minFeature = 2000,
  minCount = 1000,
  percent.mt = 20,
  method = "doubletfinder",
  cores = 1,
  save_path = NULL
)
}
\arguments{
\item{sce}{default is null, a seurat object}

\item{eset}{expression set, default is null}

\item{file_type}{if file is from 10X, data_path must be provide}

\item{data_path}{data_path of 10X data}

\item{project}{project name}

\item{index}{folder name of save_path}

\item{propotion}{default is 0.025}

\item{check_data}{if true}

\item{filter_data}{if TRUE, cells will be filtered by aforementioned criterion}

\item{minFeature}{minimum feature of cell}

\item{minCount}{minimum count of cell}

\item{percent.mt}{max of percentage of mitochondrial gens}

\item{method}{default is `doubletfinder`: Benchmarking Computational Doublet-Detection Methods for Single-Cell RNA Sequencing Data}

\item{cores}{default is 1}

\item{save_path}{default is NULL}
}
\description{
Integration of  doublet detection
}
