% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_best_resolution.R
\name{find_best_resolution}
\alias{find_best_resolution}
\title{Title}
\usage{
find_best_resolution(
  sce,
  from = 0.2,
  to = 1.2,
  save_plot = T,
  path = NULL,
  palette = "set2",
  index = NULL,
  alpha = 0.8,
  verbose = TRUE,
  return_sce = TRUE
)
}
\arguments{
\item{return_sce}{}
}
\description{
Title
}
