% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_subcluster_gene.R
\name{find_subcluster_gene}
\alias{find_subcluster_gene}
\title{Title}
\usage{
find_subcluster_gene(
  sce,
  assay = "integrated",
  slot = "scale.data",
  col_celltype = "celltypes",
  col_sub_celltype = "scpred_seurat",
  feas = NULL,
  cols = "normal",
  seed = 123,
  palette = 1,
  show_col = F,
  min_cell_count = 30,
  no_limitation_celltypes = "Epithelial cells",
  index = 1,
  path = "4-Diff-degs-of-subclusters",
  show_feas_pheatmap = 8,
  show_feas_dim = 12,
  remove_other_celltypes = TRUE,
  logfc.threshold = 0.15,
  fig.type = "pdf",
  pt.size = 0.7,
  recluster = F,
  assay_for_recluster = "integrated"
)
}
\arguments{
\item{assay_for_recluster}{}
}
\description{
Title
}
\examples{
如果目标变量没有亚组直接跳过此次循环====================
}
