% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_subcluster_signatures.R
\name{find_subcluster_signatures}
\alias{find_subcluster_signatures}
\title{Title}
\usage{
find_subcluster_signatures(
  sce,
  assay = "integrated",
  col_celltype = "Model1_merge_no_rejection",
  col_sub_celltype = NULL,
  groups = c("Model1_merge_subcluster", "integrated_snn_res.1"),
  cols = "normal",
  seed = 123,
  palette = 1,
  show_col = F,
  signature_for_deg = signature_collection,
  signature_character_limit = 60,
  min_cell_count = 50,
  no_limitation_celltypes = "Epithelial cells",
  index = NULL,
  sig_methods = c("PCAscore", "ssgsea", "AUCell"),
  path = "4-Diff-sig-of-subclusters",
  sig_file_name = "mysig",
  show_feas_pheatmap = 8,
  character_limit_heatmap = 50,
  remove_other_celltypes = TRUE,
  find_cluster_sig = FALSE,
  groups_for_cluster = NULL,
  recluster = FALSE
)
}
\arguments{
\item{sce}{seurat object}

\item{assay}{default is NULL}

\item{col_celltype}{name of celltype}

\item{col_sub_celltype}{name of subcluster}

\item{groups}{column of}

\item{cols}{default is normal}

\item{show_col}{}

\item{signature_for_deg}{signature list}

\item{signature_character_limit}{remove signature with large names}

\item{min_cell_count}{minimal cell count in each cluster}

\item{no_limitation_celltypes}{celltypes with no cell count limitation}

\item{index}{default is null, user can choose specific cell types to save computing source}

\item{sig_methods}{method used to etimate signature score}

\item{path}{path to deposided results}

\item{sig_file_name}{prefix of path}

\item{find_cluster_sig}{default is FALSE}
}
\description{
Title
}
\examples{
如果目标变量没有亚组直接跳过此次循环====================
如果目标变量没有亚组直接跳过此次循环====================
如果目标变量没有亚组直接跳过此次循环====================
如果目标变量没有亚组直接跳过此次循环====================
数据可视化==============================
如果目标变量没有亚组直接跳过此次循环====================
}
