% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_subclusters.R
\name{find_subclusters}
\alias{find_subclusters}
\title{Title}
\usage{
find_subclusters(
  sce,
  celltype,
  combined_subclusters = "celltype_subcluster",
  graph.name = NULL,
  resolution = 0.3,
  delete_new_variable = T,
  filtering_mini_cell = TRUE,
  min_cell_count = 50,
  ignore_cell_prefix = "null"
)
}
\arguments{
\item{ignore_cell_prefix}{}
}
\description{
Title
}
\examples{
批量寻找个细胞大类的亚组细胞=================================
将以上细胞亚组整合到一个列中======combined_column==================================
去掉一些离散的细胞
}
