% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_msigdb_gs.R
\name{get_msigdb_gs}
\alias{get_msigdb_gs}
\title{Title}
\usage{
get_msigdb_gs(
  msigdb = NULL,
  org = "hsa",
  category = NULL,
  subcategory = NULL,
  format = "list"
)
}
\arguments{
\item{msigdb}{default is NULL}

\item{org}{default is 'hsa', 'mus'}

\item{category}{default is NULL, other options: C1 - C8,}

\item{subcategory}{default is NULL, options:}

\item{format}{default is list}
}
\description{
Title
}
\examples{
gs<- get_msigdb_gs()
}
