% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scsig_cell_anno.R
\name{ggplotConfusionMatrix}
\alias{ggplotConfusionMatrix}
\title{draw confusion matrix}
\usage{
ggplotConfusionMatrix(input, x, y, axis_angle = 60)
}
\arguments{
\item{axis_angle}{}
}
\description{
draw confusion matrix
}
