% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheatmap_average.R
\name{pheatmap_average}
\alias{pheatmap_average}
\title{Title}
\usage{
pheatmap_average(
  sce,
  assay,
  slot,
  marker_res,
  top_n = 6,
  group = NULL,
  show_features = NULL,
  character_limit = 30,
  path = NULL,
  cols = "random",
  seed = 54321,
  show_col = T,
  palette = 1,
  fig.type = "pdf",
  width = 13,
  height = NULL,
  file_name_prefix = "0"
)
}
\arguments{
\item{palette}{}
}
\description{
Title
}
