% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sctyper_anno.R
\name{sctyper_anno}
\alias{sctyper_anno}
\title{Cell type annotation by scType}
\usage{
sctyper_anno(
  sce,
  gs_list_positive = NULL,
  gs_list_netative = NULL,
  gs_data = NULL,
  tissue_type = NULL,
  cell_type = "base",
  cell_subset = NULL,
  study = NULL,
  assay = NULL,
  slot = "scale.data",
  scale = NULL,
  cluster = "seurat_clusters",
  db_ = "ScTyperDB-merged.xlsx",
  db_path = NULL,
  gene_names_to_uppercase = TRUE
)
}
\arguments{
\item{sce}{Seurat object}

\item{gs_list_positive}{gene set list that up regulated in cell types}

\item{gs_list_netative}{gene set list that down regulated in cell types}

\item{gs_data}{gene signature data with data frame format}

\item{tissue_type}{Default is NULL}

\item{cell_type}{Cell types options: "base", "epithelial", "myeloid", "tcell", "bcell", "fibroblast" and "endothelial", default is "base"}

\item{cell_subset}{Default is NULL}

\item{study}{an option to choose gene signatures}

\item{assay}{Assay to pull from, e.g. RNA, SCT, integrated}

\item{slot}{Data slot to use, choose from 'counts', 'data', or 'scale.data'}

\item{scale}{Whether the matrix is scaled，default is NULL}

\item{cluster}{A vector of variables to group cells by}

\item{db_}{Database of manually collected cell type annotation, default is "ScTypeDB_full.xlsx" deposited in data}

\item{db_path}{path of signature data, an example: paste0(base::system.file("data", package = "scsig"),"/ScTyperDB-merged.xlsx")}

\item{gene_names_to_uppercase}{if TRUE, all gene symbol will be in uppercase}
}
\value{
Seurat object with updated metadata containing cell type annotation
}
\description{
Cell type annotation using specific marker combinations from single-cell transcriptomic data
refer to [ScType](https://www.nature.com/articles/s41467-022-28803-w)
}
\examples{
tnbc<-sctyper_anno(sce = sce, tissue = "Immune system")
}
