% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_sc.R
\name{standard_sc}
\alias{standard_sc}
\title{standard_Seurat_data_processing}
\usage{
standard_sc(
  eset = NULL,
  sce_object = NULL,
  assay = "RNA",
  file_type = "10X",
  data_path = NULL,
  project = NULL,
  qc_identity = NULL,
  nPCs = 30,
  res = 1,
  verbose = FALSE,
  index = 1,
  plot = FALSE,
  vars_dim = "Phase",
  minFeature = 2000,
  minCount = 1000,
  species = "human",
  cutoff_percent_mt = 25,
  filter_min_cells = FALSE,
  min_cell_count = 20,
  findmarkers = FALSE,
  cols = "normal",
  palette = 1,
  show_col = T,
  seed = 123,
  already_normalized = FALSE,
  save_path = NULL,
  save_sce_object = TRUE
)
}
\arguments{
\item{eset}{raw data or normalized data}

\item{sce_object}{Seurat object}

\item{assay}{default is RNA}

\item{file_type}{if file_type = "10X", data_path must be provided.}

\item{data_path}{absolute path of 10X file}

\item{project}{project name}

\item{qc_identity}{default is null}

\item{nPCs}{maximum PC of input}

\item{res}{resolution of t-SNE}

\item{verbose}{verbose}

\item{index}{number of folder name}

\item{plot}{if true, relevant plot will be drawn}

\item{vars_dim}{default is null}

\item{minFeature}{minimum features of cell}

\item{minCount}{minimum count of cell}

\item{species}{default is human}

\item{cutoff_percent_mt}{default is 25}

\item{filter_min_cells}{default is FALSE}

\item{min_cell_count}{default is 20}

\item{findmarkers}{if true, top 6 marker genes of each cluster will be defined and drawn}

\item{cols}{default is 'normal'}

\item{palette}{default is 1, other options: 2,3,4}

\item{show_col}{default is TRUE, palettes will be displayed}

\item{seed}{default is 123}

\item{already_normalized}{logistic vector, if true, `sce` object will not be normalized}

\item{save_path}{default is null}

\item{save_sce_object}{default is TRUE}
}
\description{
standard_Seurat_data_processing
}
\examples{
查看降维数据与细胞注释的关系--PCA--TSNE---UMAP------
选择需要分析的变量------labels不多的-------------------------
使用双重循环进行-------------------------
}
