% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training_sc_anno.R
\name{training_sc_anno}
\alias{training_sc_anno}
\title{Title}
\usage{
training_sc_anno(
  sce,
  group,
  assay = "RNA",
  recorrect = FALSE,
  remove_cell = NULL,
  subset_cell = FALSE,
  propotion = 0.2,
  dims = 45,
  model = "mda",
  cores = 1,
  verbose = FALSE,
  fig.type = "pdf",
  pt.size = 1,
  cols = "normal",
  seed = 123,
  show_col = F,
  palette = 4,
  width = NULL,
  height = NULL,
  show_plot = T,
  path = NULL
)
}
\arguments{
\item{recorrect}{}
}
\description{
Title
}
\examples{
利用变异的feature跑PCA
harmony会继续利用PCA进行校正
训练和初步测试预测效果
并行运算
将预测模型提取出来-这样就可以丢弃sce对象
}
